requirements.txt
streamlit>=1.32.0
groq>=0.9.0
google-generativeai>=0.7.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
langchain-text-splitters>=0.2.0
openai-whisper>=20231117
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
soundfile>=0.12.1
